requirements.txt
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
fonttools==3.42.0
idna==2.8
jieba==0.39
kiwisolver==1.1.0
lxml==4.3.4
matplotlib==3.1.0
numpy==1.16.4
Pillow==6.0.0
pymongo==3.8.0
PyMySQL==0.9.3
pyparsing==2.4.0
python-dateutil==2.8.0
redis==3.2.1
requests==2.22.0
scipy==1.3.0
six==1.12.0
tesserocr==2.4.0
urllib3==1.25.3
wordcloud==1.5.0
